services/backend/pyproject.toml
[project]
name = "igen-shop-assistant-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "paho-mqtt>=2.1",
  "psycopg2-binary",
  "sqlalchemy>=2.0",
  "pyserial>=3.5"
]

[tool.uvicorn]
reload = false
